requirements.txt
kafka-python
avro-python3
six